- adding area lore to ability list - handling aura finding actions in terms of area lore and magical Abilities - added a skeleton for the founders to be hard-coded magi to prime the system with

+++ OrderOfWizardMonks/Covenant.cs
@@ -13,7 +13,7 @@ namespace WizardMonks
 		protected Dictionary<Ability, double> _visSources;
         protected Dictionary<Ability, double> _visStock;
 		protected List<IBook> _library;
-        public int Aura { get; set; }
+        public double Aura { get; set; }
         public string Name { get; set; }
 
         public Covenant()
